Improved styling for desktop tutorial section
