+++ lib/widgets/mobile_tutorial.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+// import 'package:animate_do/animate_do.dart';
+// import '../screens/tutorial_pages.dart';
+
+class MobileTutorial extends StatelessWidget {
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      // Add your mobile tutorial layout here
+      body: Center(
+        child: Text("Mobile Tutorial"),
+      ),
+    );
+  }
+}
